исправление примера внешней задачи

+++ samples/OikTask/TmStartup.cs
@@ -22,7 +22,7 @@ namespace OikTask
     private static int              _tmCid;
     private static TmUserInfo       _userInfo;
     private static TmServerFeatures _serverFeatures;
-    private static uint             _stopEventHandle;
+    private static IntPtr           _stopEventHandle;
 
     private readonly IHostApplicationLifetime _applicationLifetime;
     private readonly ICommonInfrastructure    _infr;
